Use ASGITransport to fix isolation test failures
In updating the asgi-csrf Debian package, the new cookie tests fail due
to a ConnectError from httpx when run in an isolated environment without
internet access.

This commit updates the tests to use httpx.ASGITransport explicitly,
addressing issues related to deprecated usage of the 'app' parameter.
This change ensures that tests do not attempt to connect to external
services and aligns with new httpx practices.

Additionally, this resolves a DeprecationWarning from httpx concerning
the transition to explicit transport specification.